requirements.txt
flask
Flask-SQLAlchemy
flask-login

pyserial
qrcode[pil]